Extract ABSStore to zarr._storage.absstore (#781)
* Extract ABSStore to zarr._storage.absstore

First step towards being able to assign CODEOWNERS
to simplify maintenance. This does not yet attempt
the rename to zarr.storage.absstore in order lower
conflicts with open PRs. Imports should remain un-
changed for the moment.

see: #764

* Ignore unused import

* Register absstore.py in CODEOWNERS
