+++ .github/CODEOWNERS
@@ -1,0 +1,1 @@
+zarr/_storage/absstore.py   @zarr-developers/azure-team
